+++ packages/s2-core/src/common/constant/interaction.ts
@@ -8,6 +8,7 @@ export enum InteractionName {
   COL_ROW_RESIZE = 'rowColResize',
   DATA_CELL_MULTI_SELECTION = 'dataCellMultiSelection',
   COL_ROW_SHIFT_MULTI_SELECTION = 'colRowShiftMultiSelection',
+  SELECTED_CELL_MOVE = 'selectedCellMove',
 }
 
 export enum InteractionStateName {
@@ -62,6 +63,10 @@ export enum InteractionKeyboardKey {
   COPY = 'c',
   ESC = 'Escape',
   META = 'Meta',
+  ARROW_UP = 'ArrowUp',
+  ARROW_DOWN = 'ArrowDown',
+  ARROW_LEFT = 'ArrowLeft',
+  ARROW_RIGHT = 'ArrowRight',
 }
 
 export enum SortMethodType {
@@ -87,3 +92,8 @@ export const BRUSH_AUTO_SCROLL_INITIAL_CONFIG = {
     scroll: false,
   },
 };
+
+export enum ScrollbarPositionType {
+  CONTENT = 'content',
+  CANVAS = 'canvas',
+}