+++ packages/s2-core/src/interaction/base-interaction/hover.ts
@@ -117,6 +117,7 @@ export class HoverEvent extends BaseEvent implements BaseEventImplement {
         enterable: true,
         hideSummary: true,
         showSingleTips,
+        enableFormat: this.spreadsheet.isPivotMode(),
       };
       const data = this.getCellInfo(meta, showSingleTips);
       this.spreadsheet.showTooltipWithInfo(event, data, options);