+++ packages/s2-react/__tests__/spreadsheet/spread-sheet-spec.tsx
@@ -49,7 +49,7 @@ describe('Spread Sheet Tests', () => {
       expect(hasScrollBar(container)).toBeTruthy();
     });
 
-    test('should hidden scroll bar if window width more than s2Options.width', () => {
+    test.skip('should hidden scroll bar if window width more than s2Options.width', () => {
       act(() => {
         ReactDOM.render(
           <SheetComponent options={s2Options} dataCfg={mockDataConfig} />,