+++ s2-site/examples/react-component/drill-dwon/demo/basic-panel.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { DrillDown } from '@antv/s2';
+import { DrillDown } from '@antv/s2-react';
 
 const disabledFields = ['name'];
 const clearButtonText = '清除';